feat: login workaround (#145)
Resolves an issue where the login command would wait indefinitely or not
at all for input while waiting for the user to authenticate with RTM
before requesting the user token to save to the `~/.rtm.json` file.

Resolves #68  #131 

Many thanks to @Hostuu @dogmaticwonder @franciscof5 for letting me know
this issue was affecting more than just Chromebook users and for
patience while I resolve it!

---------

Co-authored-by: Beau Raines <beauraines@gmail.com>